Refactor schema generation
